requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
cryptography==45.0.7
cssselect2==0.8.0
distro==1.9.0
Django==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.0.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==6.0.1
Markdown==3.6
mccabe==0.7.0
openai==1.97.0
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
PyJWT==2.10.1
pypdf==6.0.0
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.0.9
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.15
